[JBIDE-13921] app lookup and creation are case insensitive (names are)
Added test for IDomain#getApplicationByName that assert it looks
applications up regardless of upper-/lowercase naming. Also added
integration tests that assert one cannot create application twice by
just changing upper-/lower-case.
