+++ src/test/java/com/openshift/internal/client/DomainResourceTest.java
@@ -11,9 +11,12 @@
 package com.openshift.internal.client;
 
 import static com.openshift.client.utils.MockUtils.anyForm;
+import static com.openshift.client.utils.Samples.ADD_APPLICATION_JSON;
 import static com.openshift.client.utils.Samples.ADD_DOMAIN_JSON;
 import static com.openshift.client.utils.Samples.DELETE_DOMAIN_JSON;
+import static com.openshift.client.utils.Samples.GET_APPLICATIONS_WITH1APP_JSON;
 import static com.openshift.client.utils.Samples.GET_APPLICATIONS_WITH2APPS_JSON;
+import static com.openshift.client.utils.Samples.GET_APPLICATIONS_WITHNOAPP_JSON;
 import static com.openshift.client.utils.Samples.GET_DOMAIN;
 import static com.openshift.client.utils.Samples.GET_DOMAINS_1EXISTING;
 import static com.openshift.client.utils.Samples.GET_DOMAINS_NOEXISTING_JSON;
@@ -28,6 +31,9 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.net.SocketTimeoutException;
 import java.net.URL;
 import java.util.List;
 
@@ -38,16 +44,22 @@ import org.junit.Test;
 import org.junit.rules.ErrorCollector;
 import org.junit.rules.ExpectedException;
 
+import com.openshift.client.ApplicationScale;
+import com.openshift.client.IApplication;
 import com.openshift.client.IDomain;
 import com.openshift.client.IGearProfile;
 import com.openshift.client.IHttpClient;
 import com.openshift.client.IOpenShiftConnection;
 import com.openshift.client.IUser;
+import com.openshift.client.InvalidCredentialsOpenShiftException;
 import com.openshift.client.OpenShiftConnectionFactory;
 import com.openshift.client.OpenShiftEndpointException;
 import com.openshift.client.OpenShiftException;
+import com.openshift.client.cartridge.IStandaloneCartridge;
 import com.openshift.client.utils.Samples;
 import com.openshift.internal.client.httpclient.BadRequestException;
+import com.openshift.internal.client.httpclient.HttpClientException;
+import com.openshift.internal.client.httpclient.UnauthorizedException;
 
 /**
  * @author Xavier Coulon
@@ -56,6 +68,7 @@ import com.openshift.internal.client.httpclient.BadRequestException;
 public class DomainResourceTest {
 
 	private IUser user;
+	private IDomain domain;
 	private IHttpClient mockClient;
 
 	@Rule
@@ -66,32 +79,42 @@ public class DomainResourceTest {
 
 	@Before
 	public void setup() throws Throwable {
-		mockClient = mock(IHttpClient.class);
+		this.mockClient = createMockClient(GET_DOMAINS_1EXISTING);
+		this.user = createUser(mockClient);
+		this.domain = user.getDomain("foobar");
+	}
+
+	private IHttpClient createMockClient(Samples domainsResponse) throws SocketTimeoutException, Throwable {
+		IHttpClient mockClient = mock(IHttpClient.class);
 		when(mockClient.get(urlEndsWith("/broker/rest/api")))
 				.thenReturn(Samples.GET_REST_API_JSON.getContentAsString());
 		when(mockClient.get(urlEndsWith("/user"))).thenReturn(Samples.GET_USER_JSON.getContentAsString());
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_1EXISTING.getContentAsString());
+		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(domainsResponse.getContentAsString());
+		return mockClient;
+	}
+
+	private IUser createUser(IHttpClient client) throws FileNotFoundException, IOException {
 		final IOpenShiftConnection connection = new OpenShiftConnectionFactory().getConnection(new RestService(
-				"http://mock", "clientId", mockClient), "foo@redhat.com", "bar");
-		this.user = connection.getUser();
+				"http://mock", "clientId", client), "foo@redhat.com", "bar");
+		return connection.getUser();
 	}
 
 	@Test
 	public void shouldLoadEmptyListOfDomains() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_NOEXISTING_JSON.getContentAsString());
+		IHttpClient client = createMockClient(GET_DOMAINS_NOEXISTING_JSON);
+		IUser user = createUser(client);
 		// operation
 		final List<IDomain> domains = user.getDomains();
 		// verifications
 		assertThat(domains).hasSize(0);
 		// 3 calls: /API + /API/user + /API/domains
-		verify(mockClient, times(3)).get(any(URL.class));
+		verify(client, times(3)).get(any(URL.class));
 	}
 
 	@Test
 	public void shouldLoadSingleUserDomain() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_1EXISTING.getContentAsString());
 		// operation
 		final List<IDomain> domains = user.getDomains();
 		// verifications
@@ -103,7 +126,6 @@ public class DomainResourceTest {
 	@Test
 	public void shouldCreateNewDomain() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_NOEXISTING_JSON.getContentAsString());
 		when(mockClient.post(anyMapOf(String.class, Object.class), urlEndsWith("/domains"))).thenReturn(
 				ADD_DOMAIN_JSON.getContentAsString());
 		// operation
@@ -116,7 +138,6 @@ public class DomainResourceTest {
 	@Test(expected = OpenShiftException.class)
 	public void shouldNotRecreateExistingDomain() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_1EXISTING.getContentAsString());
 		when(mockClient.post(anyMapOf(String.class, Object.class), urlEndsWith("/domains"))).thenReturn(
 				ADD_DOMAIN_JSON.getContentAsString());
 		// operation
@@ -128,7 +149,6 @@ public class DomainResourceTest {
 	@Test
 	public void shouldDestroyDomain() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_1EXISTING.getContentAsString());
 		when(mockClient.delete(anyForm(), urlEndsWith("/domains/foobar"))).thenReturn(
 				DELETE_DOMAIN_JSON.getContentAsString());
 		// operation
@@ -142,7 +162,6 @@ public class DomainResourceTest {
 	@Test
 	public void shouldNotDestroyDomainWithApp() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_1EXISTING.getContentAsString());
 		final BadRequestException badRequestException = new BadRequestException(
 				"Domain contains applications. Delete applications first or set force to true.", null);
 		when(mockClient.delete(anyForm(), urlEndsWith("/domains/foobar"))).thenThrow(badRequestException);
@@ -162,7 +181,6 @@ public class DomainResourceTest {
 	@Test
 	public void shouldUpdateDomainId() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_1EXISTING.getContentAsString());
 		when(mockClient.put(anyMapOf(String.class, Object.class), urlEndsWith("/domains/foobar"))).thenReturn(
 				UPDATE_DOMAIN_ID.getContentAsString());
 		final IDomain domain = user.getDomain("foobar");
@@ -178,15 +196,14 @@ public class DomainResourceTest {
 	@Test
 	public void shouldListAvailableGearSizes() throws Throwable {
 		// pre-conditions
-		when(mockClient.get(urlEndsWith("/domains"))).thenReturn(GET_DOMAINS_1EXISTING.getContentAsString());
 		when(mockClient.put(anyMapOf(String.class, Object.class), urlEndsWith("/domains/foobar"))).thenReturn(
 				UPDATE_DOMAIN_ID.getContentAsString());
 		final IDomain domain = user.getDomain("foobar");
 		// operation
 		List<IGearProfile> availableGearSizes = domain.getAvailableGearProfiles();
 		// verifications
-		assertThat(availableGearSizes).onProperty("name").contains("small", "micro", "medium", "large", "exlarge",
-				"jumbo");
+		assertThat(availableGearSizes).onProperty("name").contains(
+				"small", "micro", "medium", "large", "exlarge","jumbo");
 	}
 
 	@Test
@@ -218,6 +235,142 @@ public class DomainResourceTest {
 		verify(mockClient, times(0)).get(urlEndsWith("/domains/foobar/applications")); // no call, neither before and while refresh
 	}
 
+	@Test
+	public void shouldLoadListOfApplicationsWithNoElement() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITHNOAPP_JSON.getContentAsString());
+		// operation
+		final List<IApplication> apps = domain.getApplications();
+		// verifications
+		assertThat(apps).isEmpty();
+		// 4 calls: /API + /API/user + /API/domains +
+		// /API/domains/foobar/applications
+		verify(mockClient, times(4)).get(any(URL.class));
+	}
+
+	@Test
+	public void shouldLoadListOfApplicationsWith1Element() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITH1APP_JSON.getContentAsString());
+		// operation
+		final List<IApplication> apps = domain.getApplications();
+		// verifications
+		assertThat(apps).hasSize(1);
+		// 4 calls: /API + /API/user + /API/domains +
+		// /API/domains/foobar/applications
+		verify(mockClient, times(4)).get(any(URL.class));
+
+	}
+
+	@Test
+	public void shouldLoadListOfApplicationsWith2Elements() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITH2APPS_JSON.getContentAsString());
+		// operation
+		final List<IApplication> apps = domain.getApplications();
+		// verifications
+		assertThat(apps).hasSize(2);
+		// 4 calls: /API + /API/user + /API/domains +
+		// /API/domains/foobar/applications
+		verify(mockClient, times(4)).get(any(URL.class));
+	}
+
+	@Test(expected = InvalidCredentialsOpenShiftException.class)
+	public void shouldNotLoadListOfApplicationsWithInvalidCredentials() throws OpenShiftException,
+			HttpClientException, SocketTimeoutException {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenThrow(
+				new UnauthorizedException("invalid credentials (mock)", null));
+		// operation
+		domain.getApplications();
+		// verifications
+		// expect an exception
+	}
+
+	@Test
+	public void shouldCreateApplication() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITHNOAPP_JSON.getContentAsString());
+		when(mockClient.post(anyForm(), urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				ADD_APPLICATION_JSON.getContentAsString());
+		// operation
+		final IStandaloneCartridge cartridge = new StandaloneCartridge("jbossas-7");
+		final IApplication app = domain.createApplication("sample", cartridge, ApplicationScale.NO_SCALE, null);
+		// verifications
+		assertThat(app.getName()).isEqualTo("sample");
+		assertThat(app.getGearProfile().getName()).isEqualTo("small");
+		assertThat(app.getApplicationScale()).isEqualTo(ApplicationScale.NO_SCALE);
+		assertThat(app.getApplicationUrl()).isEqualTo("http://sample-foobar.stg.rhcloud.com/");
+		assertThat(app.getCreationTime()).isNotNull();
+		assertThat(app.getGitUrl()).isNotNull().startsWith("ssh://")
+				.endsWith("@sample-foobar.stg.rhcloud.com/~/git/sample.git/");
+		assertThat(app.getCartridge()).isEqualTo(cartridge);
+		assertThat(app.getUUID()).isNotNull();
+		assertThat(app.getDomain()).isEqualTo(domain);
+		assertThat(LinkRetriever.retrieveLinks(app)).hasSize(14);
+		assertThat(domain.getApplications()).hasSize(1).contains(app);
+	}
+
+	@Test(expected = OpenShiftException.class)
+	public void shouldNotCreateApplicationWithMissingName() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITH2APPS_JSON.getContentAsString());
+		when(mockClient.post(anyForm(), urlEndsWith("/domains"))).thenReturn(ADD_DOMAIN_JSON.getContentAsString());
+		// operation
+		domain.createApplication(null, new StandaloneCartridge("jbossas-7"), null, null);
+		// verifications
+		// expected exception
+	}
+
+	@Test(expected = OpenShiftException.class)
+	public void shouldNotCreateApplicationWithMissingCartridge() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITH2APPS_JSON.getContentAsString());
+		when(mockClient.post(anyForm(), urlEndsWith("/domains"))).thenReturn(ADD_DOMAIN_JSON.getContentAsString());
+		// operation
+		domain.createApplication("foo", null, null, null);
+		// verifications
+		// expected exception
+	}
+
+	@Test
+	public void shouldNotRecreateExistingApplication() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITH2APPS_JSON.getContentAsString());
+		// operation
+		try {
+			domain.createApplication("sample", new StandaloneCartridge("jbossas-7"), null, null);
+			// expect an exception
+			fail("Expected exception here...");
+		} catch (OpenShiftException e) {
+			// OK
+		}
+		// verifications
+		assertThat(domain.getApplications()).hasSize(2);
+	}
+
+	@Test
+	public void shouldGetApplicationByNameCaseInsensitive() throws Throwable {
+		// pre-conditions
+		when(mockClient.get(urlEndsWith("/domains/foobar/applications"))).thenReturn(
+				GET_APPLICATIONS_WITH2APPS_JSON.getContentAsString());
+		// operation
+		IApplication lowerCaseQueryResult = domain.getApplicationByName("scalable");
+		IApplication upperCaseQueryResult = domain.getApplicationByName("SCALABLE");
+
+		// verifications
+		assertThat(lowerCaseQueryResult).isNotNull();
+		assertThat(lowerCaseQueryResult.getName()).isEqualTo("scalable");
+		assertThat(upperCaseQueryResult).isNotNull();
+		assertThat(upperCaseQueryResult.getName()).isEqualTo("scalable");
+	}
 	
 	@Test
 	@Ignore
